Move supplementary evidence to strategy
